+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+# 1.0.0
+
+- (feature) Add configuration
+
 # 0.0.1
 
-*   (feature) Storage controller
-*   (feature) Transactions
-*   (feature) Embedable struct for reepositories
+- (feature) Storage controller
+- (feature) Transactions
+- (feature) Embedable struct for reepositories